requirements.txt
asgiref==3.3.1
auditlog3==1.0.1
authy==2.2.6
braintree==4.5.0
cachetools==4.2.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-admin-interface==0.14.2
django-adminlte3==0.1.6
django-allauth==0.44.0
django-braces==1.14.0
django-colorfield==0.3.2
django-excel==0.0.10
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==2.14.3
django-hosts==4.0
django-import-export==2.5.0
django-jet==1.0.8
django-jet-new==2.2.8
django-oauth-toolkit==1.3.3
django-payments==0.13.0
django-payments-razorpay==1.0
django-payu==0.5
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-jwt==0.1.4
djangorestframework-simplejwt==4.6.0
ecdsa==0.14.1
et-xmlfile==1.0.1
google-api-core==1.24.1
google-api-python-client==1.4.1
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
lml==0.1.0
MarkupPy==1.14
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
PyJWT==2.0.1
pyparsing==2.4.7
python-dotenv==0.15.0
python-jose==3.2.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stripe==2.55.1
suds-jurko==0.6
tablib==3.0.0
texttable==1.6.3
twilio==6.48.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
